pyproject.toml
[project]
name = "rag-platform"
version = "0.1.0"
description = "Multi-tenant enterprise RAG platform"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "asyncpg>=0.30",
    "qdrant-client>=1.12",
    "sentence-transformers>=3.3",
    "groq>=0.13",
    "google-generativeai>=0.8",
    "pypdf>=5.1",
    "python-docx>=1.1",
    "tenacity>=9.0",
    "typer>=0.13",
    "python-multipart>=0.0.20",
    "httpx>=0.28",
    "structlog>=24.4",
]

[project.scripts]
rag = "app.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
